Better. See #46.
